Remove useless casts

+++ src/tests/unit-tests/signal-processing-blocks/adapter/adapter_test.cc
@@ -373,9 +373,9 @@ TEST_F(DataTypeAdapter, IshortToCshortValidationOfResults)
         {
             while (ifs.read(reinterpret_cast<char*>(&iSample), sizeof(lv_16sc_t)))
                 {
-                    EXPECT_EQ(input_data_shorts.at(i), static_cast<int16_t>(iSample.real()));
+                    EXPECT_EQ(input_data_shorts.at(i), iSample.real());
                     i++;
-                    EXPECT_EQ(input_data_shorts.at(i), static_cast<int16_t>(iSample.imag()));
+                    EXPECT_EQ(input_data_shorts.at(i), iSample.imag());
                     i++;
                 }
         }
